CH1_Introduction/Stagenography/requirements.txt
pillow
auto_py_to_exe